requirements.txt
python=3.7
pip
numpy
pandas
PyMySQL=0.10.1
PyYAML
SQLAlchemy=1.3.23
xmltodict
